Added picture-in-picture
Added support for specifications of "Max" to support picture-in-picture,
where one screen is max-size and others are small.  Added some pip
window arrangements to the settings file.

Also added "Full", just because it might come in handy, although it's
sort of against the purpose of this project because you can't see any
other windows when one is fullscreen.