Fix bug when using swap with pip
When swapping windows, if there was more than one picture-in-picture
window, one of them would disappear. This is because the swapped windows
would both be put in front of the other windows. So this version goes
through the windows and touches each of them in numerical order, to make
sure that it gets set back to having the little (greater-numbered)
windows on top.