Added Reserve Area feature. Tweaked some window focus for picture-in-picture.
Added a feature to set a 'reserve area' around any or all edges of the
screen, where no kodis will be placed. This is for if you want to put
other things there, especially CPU or network monitors, or if you just
wanted to use part of your screen for other purposes. See entries in
settings file.

Changed how it handles window focus when swapping, so that it correctly
handles picture-in-picture cases, where the previous implementation
would end up with the little window behind the big one.

Apparently fixed a bug with $current_kodis and $num_kodis, but I don't
remember the details...