Added ability to run shell scripts
Added a feature wherein a line that starts with "> " (space optional)
will be run as a shell command. This is to make it easier to perform
automated operations interleaved with remote commands, so you don't have
to have an outside shell script, and/or multiple key scripts, to perform
operations that involve keystrokes and other commands such as swaps.